Speed up historical analysis loading via filename-level dedupe and coarse progress reporting, add explicit IBKR phase status in portfolio_manager.py, and harden recent refactor seams by making tool hooks opt-in, tightening audit/runtime typing and tests, fixing consultant tool-call ID fallback, and preserving explicit private-agent imports without star-export leakage
