+++ src/agents/analyst_nodes.py
@@ -1,0 +1,411 @@
+from __future__ import annotations
+
+import re
+from collections.abc import Callable
+from datetime import datetime
+from typing import Any
+
+import structlog
+from langchain_core.messages import AIMessage, HumanMessage, SystemMessage
+from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
+from langgraph.types import RunnableConfig
+
+from . import message_utils, support
+from . import runtime as agent_runtime
+from .state import AgentState
+
+logger = structlog.get_logger(__name__)
+
+
+def create_analyst_node(
+    llm,
+    agent_key: str,
+    tools: list[Any],
+    output_field: str,
+    retry_llm: Any | None = None,
+    allow_retry: bool = False,
+) -> Callable:
+    """
+    Factory function creating data analyst agent nodes.
+    """
+
+    async def analyst_node(state: AgentState, config: RunnableConfig) -> dict[str, Any]:
+        from src.prompts import get_prompt
+
+        agent_prompt = get_prompt(agent_key)
+        if not agent_prompt:
+            logger.error("missing_prompt", agent=agent_key)
+            return {output_field: f"Error: Could not load prompt for {agent_key}."}
+
+        messages_template = [MessagesPlaceholder(variable_name="messages")]
+        prompt_template = ChatPromptTemplate.from_messages(messages_template)
+        runnable = (
+            prompt_template | llm.bind_tools(tools) if tools else prompt_template | llm
+        )
+
+        try:
+            prompts_used = state.get("prompts_used", {})
+            prompts_used[output_field] = {
+                "agent_name": agent_prompt.agent_name,
+                "version": agent_prompt.version,
+            }
+
+            filtered_messages = message_utils.filter_messages_for_gemini(
+                state.get("messages", []), agent_key=agent_key
+            )
+            msg_types = [type(message).__name__ for message in filtered_messages]
+            msg_has_tool_calls = [
+                bool(getattr(message, "tool_calls", None))
+                for message in filtered_messages
+                if hasattr(message, "tool_calls")
+            ]
+            logger.debug(
+                "analyst_filtered_messages",
+                agent_key=agent_key,
+                total_state_messages=len(state.get("messages", [])),
+                filtered_count=len(filtered_messages),
+                message_types=msg_types,
+                has_tool_calls_list=msg_has_tool_calls,
+            )
+
+            context = support.get_context_from_config(config)
+            current_date = (
+                context.trade_date if context else datetime.now().strftime("%Y-%m-%d")
+            )
+            ticker = (
+                context.ticker
+                if context
+                else state.get("company_of_interest", "UNKNOWN")
+            )
+            company_name = state.get("company_name", ticker)
+            company_resolved = state.get("company_name_resolved", True)
+
+            extra_context = ""
+
+            if agent_key == "junior_fundamentals_analyst":
+                news_report = state.get("news_report", "")
+                if news_report:
+                    extra_context = (
+                        "\n\n### NEWS CONTEXT (for ADR/analyst search queries)"
+                        f"\n{news_report}\n"
+                    )
+
+            if agent_key == "fundamentals_analyst":
+                raw_data = state.get("raw_fundamentals_data", "")
+                foreign_data = state.get("foreign_language_report", "")
+                news_report = state.get("news_report", "")
+
+                if raw_data:
+                    extra_context = (
+                        "\n\n### RAW FINANCIAL DATA FROM JUNIOR ANALYST (Primary Source)"
+                        f"\n{raw_data}\n"
+                    )
+                else:
+                    logger.warning(
+                        "senior_fundamentals_no_raw_data",
+                        ticker=ticker,
+                        message="Junior Analyst data not available - this should not happen",
+                    )
+
+                if foreign_data:
+                    extra_context += (
+                        "\n\n### FOREIGN/ALTERNATIVE SOURCE DATA (Cross-Reference)"
+                        "\nNote: Use this data to fill gaps in Junior Analyst data. "
+                        "Prioritize Junior's data when both sources have the same metric.\n"
+                        f"{foreign_data}\n"
+                    )
+                    logger.info(
+                        "senior_fundamentals_has_foreign_data",
+                        ticker=ticker,
+                        foreign_data_length=len(foreign_data),
+                    )
+                else:
+                    logger.info(
+                        "senior_fundamentals_no_foreign_data",
+                        ticker=ticker,
+                        message="Foreign Language Analyst data not available - proceeding with Junior data only",
+                    )
+
+                if news_report:
+                    news_highlights = support.extract_news_highlights(news_report)
+                    extra_context += (
+                        "\n\n### NEWS HIGHLIGHTS (for Qualitative Growth Scoring)"
+                        f"\n{news_highlights}\n"
+                    )
+                else:
+                    logger.info(
+                        "senior_fundamentals_no_news",
+                        ticker=ticker,
+                        message="News report not yet available (parallel execution) - proceeding without news context",
+                    )
+
+                conflict_report = support.compute_data_conflicts(raw_data, foreign_data)
+                if conflict_report:
+                    extra_context += conflict_report
+                    logger.info(
+                        "senior_fundamentals_conflicts_detected",
+                        ticker=ticker,
+                        conflict_count=conflict_report.count("\n- "),
+                    )
+
+                legal_report = state.get("legal_report", "")
+                if legal_report:
+                    extra_context += (
+                        "\n\n### LEGAL/TAX RISK ASSESSMENT (From Legal Counsel)"
+                        "\nUse this to inform your PFIC_RISK assessment in DATA_BLOCK. "
+                        "If Legal Counsel found PFIC disclosure (pfic_status: PROBABLE), set PFIC_RISK: MEDIUM or HIGH. "
+                        "If no disclosure found in high-risk sector (pfic_status: UNCERTAIN), set PFIC_RISK: MEDIUM.\n"
+                        f"{legal_report}\n"
+                    )
+                    logger.info(
+                        "senior_fundamentals_has_legal_data",
+                        ticker=ticker,
+                        legal_data_length=len(legal_report),
+                    )
+                else:
+                    logger.info(
+                        "senior_fundamentals_no_legal_data",
+                        ticker=ticker,
+                        message="Legal Counsel data not yet available - proceeding without legal context",
+                    )
+
+            if agent_key == "news_analyst":
+                try:
+                    from src.memory import create_macro_events_store
+                    from src.retrospective import _get_ticker_suffix
+
+                    macro_store = create_macro_events_store()
+                    if macro_store.available:
+                        region = _get_ticker_suffix(ticker)
+                        events = macro_store.get_active_events(
+                            region_filter=region or None
+                        )
+                        if events:
+                            lines = ["### MACRO EVENT CONTEXT (portfolio-detected)"]
+                            for event in events[:2]:
+                                lines.append(
+                                    f"- {event.event_date} | {event.impact} | "
+                                    f"{event.scope}: {event.news_headline}"
+                                )
+                                if event.news_detail:
+                                    lines.append(f"  {event.news_detail}")
+                            lines.append(
+                                "Instruction: Determine if this equity is an "
+                                "'Innocent Bystander' (dropped due to the macro event, "
+                                "fundamentals intact -> OPPORTUNITY) or "
+                                "'Structurally Impaired' (business model affected -> EXIT). "
+                                "Ignore if event is inapplicable to this region/sector."
+                            )
+                            extra_context += "\n\n" + "\n".join(lines) + "\n"
+                            logger.info(
+                                "macro_events_injected",
+                                ticker=ticker,
+                                count=len(events[:2]),
+                            )
+                except Exception as exc:
+                    logger.debug("macro_events_injection_failed", error=str(exc))
+
+            full_system_instruction = (
+                f"{agent_prompt.system_message}\n\n"
+                f"Date: {support._format_date_with_fy_hint(current_date)}\n"
+                f"Ticker: {ticker}\n"
+                f"{support._company_line(company_name, company_resolved)}\n"
+                f"{support.get_analysis_context(ticker)}"
+                f"{extra_context}"
+            )
+            invocation_messages = [
+                SystemMessage(content=full_system_instruction),
+                *filtered_messages,
+            ]
+
+            response = await agent_runtime.invoke_with_rate_limit_handling(
+                runnable,
+                {"messages": invocation_messages},
+                context=agent_prompt.agent_name,
+            )
+            response.name = agent_key
+
+            new_state = {
+                "sender": agent_key,
+                "messages": [response],
+                "prompts_used": prompts_used,
+            }
+
+            tool_calls = getattr(response, "tool_calls", None)
+            has_tool_calls = isinstance(tool_calls, list) and len(tool_calls) > 0
+            logger.info(
+                "analyst_response_details",
+                agent_key=agent_key,
+                content_type=type(response.content).__name__,
+                content_len=len(response.content) if response.content else 0,
+                tool_calls_count=len(tool_calls) if isinstance(tool_calls, list) else 0,
+                has_tool_calls=has_tool_calls,
+            )
+
+            if has_tool_calls:
+                return new_state
+
+            content_str = message_utils.extract_string_content(response.content)
+
+            if (
+                allow_retry
+                and retry_llm is not None
+                and support._is_output_insufficient(content_str, agent_key)
+            ):
+                logger.warning(
+                    "analyst_retry_with_deep_thinking",
+                    agent_key=agent_key,
+                    ticker=ticker,
+                    original_length=len(content_str),
+                    has_datablock="DATA_BLOCK" in content_str if content_str else False,
+                    message="Insufficient output from quick LLM, retrying once with deep thinking",
+                )
+                retry_runnable = (
+                    prompt_template | retry_llm.bind_tools(tools)
+                    if tools
+                    else prompt_template | retry_llm
+                )
+
+                try:
+                    retry_response = (
+                        await agent_runtime.invoke_with_rate_limit_handling(
+                            retry_runnable,
+                            {"messages": invocation_messages},
+                            context=f"{agent_prompt.agent_name} (RETRY-HIGH)",
+                        )
+                    )
+                    retry_response.name = agent_key
+                    retry_content_str = message_utils.extract_string_content(
+                        retry_response.content
+                    )
+                    retry_tool_calls = getattr(retry_response, "tool_calls", None)
+                    retry_has_tool_calls = (
+                        isinstance(retry_tool_calls, list) and len(retry_tool_calls) > 0
+                    )
+
+                    if retry_has_tool_calls:
+                        new_state["messages"] = [retry_response]
+                        logger.info(
+                            "analyst_retry_produced_tool_calls",
+                            agent_key=agent_key,
+                            ticker=ticker,
+                        )
+                        return new_state
+
+                    logger.info(
+                        "analyst_retry_complete",
+                        agent_key=agent_key,
+                        ticker=ticker,
+                        original_length=len(content_str),
+                        retry_length=len(retry_content_str),
+                        retry_has_datablock=(
+                            "DATA_BLOCK" in retry_content_str
+                            if retry_content_str
+                            else False
+                        ),
+                        retry_improved=len(retry_content_str) > len(content_str),
+                    )
+                    content_str = retry_content_str
+                except Exception as retry_error:
+                    logger.error(
+                        "analyst_retry_failed",
+                        agent_key=agent_key,
+                        ticker=ticker,
+                        error=str(retry_error),
+                    )
+
+            new_state[output_field] = content_str
+
+            if agent_key == "fundamentals_analyst":
+                logger.info(
+                    "fundamentals_output",
+                    has_datablock="DATA_BLOCK" in content_str,
+                    length=len(content_str),
+                )
+            return new_state
+        except Exception as exc:
+            logger.error(
+                "analyst_node_error", output_field=output_field, error=str(exc)
+            )
+            return {
+                "messages": [AIMessage(content=f"Error: {str(exc)}")],
+                output_field: f"Error: {str(exc)}",
+            }
+
+    return analyst_node
+
+
+def create_valuation_calculator_node(llm) -> Callable:
+    """
+    Factory function creating Valuation Calculator node for chart generation.
+    """
+
+    async def valuation_calculator_node(
+        state: AgentState, config: RunnableConfig
+    ) -> dict[str, Any]:
+        from src.prompts import get_prompt
+
+        agent_prompt = get_prompt("valuation_calculator")
+        if not agent_prompt:
+            logger.error("missing_prompt", agent="valuation_calculator")
+            return {"valuation_params": ""}
+
+        ticker = state.get("company_of_interest", "UNKNOWN")
+        company_name = state.get("company_name", ticker)
+        fundamentals_report = state.get("fundamentals_report", "")
+
+        if not isinstance(fundamentals_report, str):
+            fundamentals_report = message_utils.extract_string_content(
+                fundamentals_report
+            )
+
+        if not fundamentals_report or "DATA_BLOCK" not in fundamentals_report:
+            logger.warning(
+                "valuation_calculator_no_datablock",
+                ticker=ticker,
+                message="No DATA_BLOCK found - skipping valuation params extraction",
+            )
+            return {"valuation_params": ""}
+
+        data_block_pattern = (
+            r"### --- START DATA_BLOCK[^\n]*---(.+?)### --- END DATA_BLOCK ---"
+        )
+        blocks = list(re.finditer(data_block_pattern, fundamentals_report, re.DOTALL))
+        if not blocks:
+            logger.warning(
+                "valuation_calculator_datablock_regex_failed",
+                ticker=ticker,
+                message="DATA_BLOCK text present but regex extraction failed",
+            )
+            return {"valuation_params": ""}
+        data_block = blocks[-1].group(0)
+
+        prompt = f"""{agent_prompt.system_message}
+
+TICKER: {ticker}
+COMPANY: {company_name}
+
+DATA_BLOCK:
+{data_block}
+
+Extract valuation parameters and output in the required format."""
+
+        try:
+            response = await agent_runtime.invoke_with_rate_limit_handling(
+                llm,
+                [HumanMessage(content=prompt)],
+                context=agent_prompt.agent_name,
+            )
+            content_str = message_utils.extract_string_content(response.content)
+            logger.info(
+                "valuation_calculator_complete",
+                ticker=ticker,
+                has_params_block="VALUATION_PARAMS" in content_str,
+                content_length=len(content_str),
+            )
+            return {"valuation_params": content_str}
+        except Exception as exc:
+            logger.error("valuation_calculator_error", ticker=ticker, error=str(exc))
+            return {"valuation_params": ""}
+
+    return valuation_calculator_node
